♻️ refactor: split agent/pool.go into focused files
- pool_options.go: PoolOption, ChatOption, all With* funcs
- pool_reaper.go: StartReaper, reap (idle/dead runner cleanup)
- pool_compaction.go: CompactSession, NeedsCompaction, compaction prompt
- pool.go: Pool struct, session CRUD, Chat, runner lifecycle, persistence
